+++ Sources/Extensions/UIKit/UIApplication/UIApplicationExtension.swift
@@ -7,6 +7,8 @@
 
 import UIKit
 
+// MARK: - ViewController
+
 extension UIApplication {
 
     public func topViewController(from baseVC: UIViewController? = UIApplication.shared.keyWindow?.rootViewController)
@@ -26,6 +28,31 @@ extension UIApplication {
     }
 
 }
+
+// MARK: - Open
+
+@available(iOS 10.0, tvOS 10.0, *)
+extension UIApplication {
+
+    /// Open app settings
+    public func openAppSettings() {
+        guard let url = URL(string: UIApplication.openSettingsURLString) else {
+            return
+        }
+        open(url, options: [:], completionHandler: nil)
+    }
+
+    /// Open app review page
+    /// - Parameter url: `URL` App page url finishing with `write-review`
+    public func openAppStoreReviewPage(_ url: URL) {
+        DispatchQueue.main.async {
+            UIApplication.shared.open(url, options: [:], completionHandler: nil)
+        }
+    }
+
+}
+
+// MARK: - AppDelegate
 
 extension UIApplication {
     public static func delegate<T: UIApplicationDelegate>(_ type: T.Type) -> T? {